Extend S3 bucket policy to ensure encryption-in-transit
Whilst the current option policy ensures server-side encryption,
encryption of the transport mechanism isn't enforced.

This change extends the S3 bucket policy to enforce encryption in
transit, which is necessary to satisfy the s3-bucket-ssl-requests-only
AWS Config Rule[1]

Given the option to prevent unencrypted uploads is already present,
and this change fits in the spirit of that, I've not introduced a
separate flag for this behaviour.

[1] https://docs.aws.amazon.com/config/latest/developerguide/s3-bucket-ssl-requests-only.html